requirements.txt
# DeepGTA - Multi-Object Tracking Pipeline
# Core dependencies

# Deep Learning
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Scientific Computing
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Linear Assignment
lap>=0.4.0
cython>=0.29.0
cython-bbox>=0.1.3

# Progress and Logging
tqdm>=4.60.0
loguru>=0.6.0

# Model Download
gdown>=4.0.0

# Optional: TorchReID (for advanced ReID features)
# Can be installed separately if needed
